Creating version module. Moving CidVersion class from within the template module to make it re-usable for other modules such as the definition module.
